[refactor] 헤더 챗리스트 - 시간 형식 변환 분리
